+++ misc/router-benchmark/README.md
@@ -11,3 +11,31 @@ Benchmark_ServeMux-7             3167156               374.2 ns/op            80
 Benchmark_gorillamux-7            760047              1560 ns/op            1312 B/op         10 allocs/op
 Benchmark_httprouter-7           3195236               371.7 ns/op           504 B/op          5 allocs/op
 ```
+
+## Lines of code
+`cloc routeServeMux.go`
+```shell
+-------------------------------------------------------------------------------
+Language                     files          blank        comment           code
+-------------------------------------------------------------------------------
+Go                               1              6              9             53
+-------------------------------------------------------------------------------
+```
+
+`cloc routeGorillamux.go`
+```shell
+-------------------------------------------------------------------------------
+Language                     files          blank        comment           code
+-------------------------------------------------------------------------------
+Go                               1              2              1             11
+-------------------------------------------------------------------------------
+```
+
+`cloc routeHttprouter.go`
+```shell
+-------------------------------------------------------------------------------
+Language                     files          blank        comment           code
+-------------------------------------------------------------------------------
+Go                               1              2              1             11
+-------------------------------------------------------------------------------
+```
